requirements.txt
Django==5.1.5
mysqlclient==2.2.4
django-sslserver==0.22
django-extensions==3.2.3
opencv-python==4.9.0.80
numpy==1.26.4
face-recognition==1.3.0
Pillow==10.2.0
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2 